+++ abacusflow-usecase/abacusflow-usecase-inventory/src/main/kotlin/org/bruwave/abacusflow/usecase/inventory/InventoryService.kt
@@ -30,8 +30,8 @@ interface InventoryService {
 
     fun adjustWarningLine(
         id: Long,
-        newSafetyStock: Int,
-        newMaxStock: Int,
+        newSafetyStock: Long,
+        newMaxStock: Long,
     )
 
     fun checkSafetyStock(id: Long): Boolean
